+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # Changelog
 
-## [Unreleased](https://github.com/jn-jairo/laravel-cast/compare/v0.0.1...master)
+## [Unreleased](https://github.com/jn-jairo/laravel-cast/compare/v0.0.2...master)
 
-## [v0.0.1 [2019-09-01]](https://github.com/jn-jairo/laravel-cast/commit/9a09ab219c984878ddfa124259e423d6b39dba48)
+## [v0.0.2 (2019-09-14)](https://github.com/jn-jairo/laravel-cast/compare/v0.0.1...v0.0.2)
+
+### Added
+- Laravel 6 support
+- Test using [orchestral/testbench](https://github.com/orchestral/testbench)
+
+### Changed
+- Minimal PHP version 7.2
+- Minimal Laravel version 5.8
+
+## [v0.0.1 (2019-09-01)](https://github.com/jn-jairo/laravel-cast/commit/9a09ab219c984878ddfa124259e423d6b39dba48)
 - Cast
     - `int`, `integer`
     - `float`, `real`, `double`